pyproject.toml
[tool.poetry]
name = "fast-auth"
version = "0.1.0"
description = "Demonstrate simple OAuth through Auth server microservice built with Fast-API"
authors = ["rifatrakib <Rakib.1508@outlook.com>"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = "^3.9"
anyio = "3.6.1"
asgiref = "3.5.2"
async-timeout = "4.0.2"
certifi = "2022.6.15"
cfgv = "3.3.1"
charset-normalizer = "2.1.1"
click = "8.1.3"
colorama = "0.4.5"
Deprecated = "1.2.13"
distlib = "0.3.6"
dnspython = "2.2.1"
email-validator = "1.2.1"
fastapi = "0.79.1"
filelock = "3.8.0"
greenlet = "1.1.2"
h11 = "0.13.0"
httptools = "0.4.0"
identify = "2.5.5"
idna = "3.3"
itsdangerous = "2.1.2"
Jinja2 = "3.1.2"
MarkupSafe = "2.1.1"
nodeenv = "1.7.0"
orjson = "3.7.12"
packaging = "21.3"
platformdirs = "2.5.2"
pre-commit = "2.20.0"
psycopg2 = "2.9.3"
pydantic = "1.9.2"
pyparsing = "3.0.9"
python-dotenv = "0.20.0"
python-multipart = "0.0.5"
PyYAML = "6.0"
redis = "4.3.4"
requests = "2.28.1"
six = "1.16.0"
sniffio = "1.2.0"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.41"}
starlette = "0.19.1"
toml = "0.10.2"
typing-extensions = "4.3.0"
ujson = "5.4.0"
urllib3 = "1.26.11"
uvicorn = "0.17.6"
virtualenv = "20.16.5"
watchgod = "0.8.2"
websockets = "10.3"
wrapt = "1.14.1"
alembic = "^1.8.1"
asyncpg = "^0.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 119
